OAP-32 Limpeza de código e ajuste de testes
