--- source/otus-business/src/main/java/br/org/otus/email/OtusEmail.java
@@ -1,23 +1,0 @@
-package br.org.otus.email;
-
-import br.org.owail.sender.email.Recipient;
-import br.org.owail.sender.email.Sender;
-
-import java.util.List;
-import java.util.Map;
-
-public interface OtusEmail {
-
-  String getTemplatePath();
-
-  Sender getFrom();
-
-  List<Recipient> getRecipients();
-
-  String getSubject();
-
-  Map<String, String> getContentDataMap();
-
-  String getContentType();
-
-}
