+++ source/otus-business/src/main/java/br/org/otus/email/service/EmailNotifierServiceBean.java
@@ -4,31 +4,21 @@ import br.org.otus.communication.CommunicationDataBuilder;
 import br.org.otus.communication.GenericCommunicationData;
 import br.org.otus.configuration.dto.OtusInitializationConfigDto;
 import br.org.otus.email.BasicEmailSender;
-import br.org.otus.email.OtusEmail;
-import br.org.otus.email.OtusEmailFactory;
-import br.org.otus.email.system.SystemInstallationEmail;
 import br.org.otus.gateway.gates.CommunicationGatewayService;
 import br.org.otus.gateway.response.exception.ReadRequestException;
 import br.org.otus.security.EncryptorResources;
 import br.org.otus.system.SystemConfigDaoBean;
 import br.org.otus.user.signup.SignupService;
-import br.org.owail.io.TemplateReader;
-import br.org.owail.sender.email.EmailCompositionException;
-import br.org.owail.sender.email.Recipient;
 import br.org.owail.sender.email.Sender;
-import br.org.owail.sender.gmail.GMailer;
 import br.org.tutty.Equalizer;
 import org.ccem.otus.exceptions.webservice.common.DataNotFoundException;
 import org.ccem.otus.exceptions.webservice.http.EmailNotificationException;
 import org.ccem.otus.exceptions.webservice.security.EncryptedException;
 
-import javax.ejb.Asynchronous;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
-import javax.mail.MessagingException;
 import javax.persistence.NoResultException;
 import java.net.MalformedURLException;
-import java.util.Map;
 import java.util.logging.Logger;
 
 @Stateless
@@ -56,31 +46,6 @@ public class EmailNotifierServiceBean implements EmailNotifierService {
     }
   }
 
-
-
-  @Override
-  @Asynchronous
-  public void sendEmail(OtusEmail email) throws EmailNotificationException {
-    sendEmailSync(email);
-  }
-
-  @Override
-  public void sendEmailSync(OtusEmail email) throws EmailNotificationException {
-    GMailer mailer = GMailer.createTLSMailer();
-
-    mailer.setFrom(email.getFrom());
-    mailer.addRecipients(email.getRecipients());
-    mailer.setSubject(email.getSubject());
-    mailer.setContentType(email.getContentType());
-    mailer.setContent(mergeTemplate(email.getContentDataMap(), email.getTemplatePath()));
-
-    try {
-      mailer.send();
-    } catch (MessagingException | EmailCompositionException e) {
-      e.printStackTrace();
-    }
-  }
-
   @Override
   public Sender getSender() throws EncryptedException, DataNotFoundException {
     try {
@@ -90,12 +55,6 @@ public class EmailNotifierServiceBean implements EmailNotifierService {
     } catch (NoResultException e) {
       throw new DataNotFoundException();
     }
-  }
-
-  private String mergeTemplate(Map<String, String> dataMap, String template) {
-    TemplateReader templateReader = new TemplateReader();
-    String templateContent = templateReader.getFileToString(getClass().getClassLoader(), template);
-    return templateReader.fillTemplate(dataMap, templateContent);
   }
 
 }
