+++ source/otus-business/src/main/java/br/org/otus/user/management/ManagementUserServiceBean.java
@@ -3,8 +3,6 @@ package br.org.otus.user.management;
 import br.org.otus.communication.CommunicationDataBuilder;
 import br.org.otus.communication.GenericCommunicationData;
 import br.org.otus.email.service.EmailNotifierService;
-import br.org.otus.email.user.management.DisableUserNotificationEmail;
-import br.org.otus.email.user.management.EnableUserNotificationEmail;
 import br.org.otus.gateway.gates.CommunicationGatewayService;
 import br.org.otus.gateway.response.exception.ReadRequestException;
 import br.org.otus.model.User;
@@ -17,7 +15,6 @@ import br.org.otus.user.signup.SignupService;
 import br.org.tutty.Equalizer;
 import org.bson.types.ObjectId;
 import org.ccem.otus.exceptions.webservice.common.DataNotFoundException;
-import org.ccem.otus.exceptions.webservice.http.EmailNotificationException;
 import org.ccem.otus.exceptions.webservice.security.EncryptedException;
 import org.ccem.otus.exceptions.webservice.validation.ValidationException;
 import org.ccem.otus.model.FieldCenter;
